third_party/blink/web_tests/xr/xrWebGLLayer_incompatible_device.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/testdriver.js"></script>
<script src="../resources/testdriver-vendor.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="../external/wpt/webxr/resources/webxr_test_constants.js"></script>
<canvas />

<script>
promise_test( (t) => {
  let webglCanvas = document.getElementsByTagName('canvas')[0];
  let glAttributes = {
    alpha: false,
    antialias: false,
  };
  let gl = webglCanvas.getContext('webgl', glAttributes);
  return navigator.xr.test.simulateDeviceConnection(TRACKED_IMMERSIVE_DEVICE)
    .then( (controller) => new Promise((resolve) => {
      navigator.xr.test.simulateUserActivation( () => {
        resolve(navigator.xr.requestSession('immersive-vr')
          .then( (session) => new Promise((resolve) => {
            try {
              let webglLayer = new XRWebGLLayer(session, gl);
              reject("XRWebGLLayer should fail with invalid arguments");
            } catch (err) {
              assert_equals(err.name, "InvalidStateError");
              resolve();
            }
        })));
      });
    }));
}, "XRWebGLLayer throws error when the session device is not the set device");

</script>
